feat: nuking everything
