--- emoji-map/Utilities/CoordinateWrapper.swift
@@ -1,42 +1,0 @@
-//
-//  CoordinateWrapper.swift
-//  emoji-map
-//
-//  Created by Enrique on 3/2/25.
-//
-
-
-import CoreLocation
-
-struct CoordinateWrapper: Equatable, Codable {
-    let coordinate: CLLocationCoordinate2D
-    
-    init(_ coordinate: CLLocationCoordinate2D) {
-        self.coordinate = coordinate
-    }
-    
-    // MARK: - Codable Implementation
-    
-    enum CodingKeys: String, CodingKey {
-        case latitude
-        case longitude
-    }
-    
-    init(from decoder: Decoder) throws {
-        let container = try decoder.container(keyedBy: CodingKeys.self)
-        let latitude = try container.decode(Double.self, forKey: .latitude)
-        let longitude = try container.decode(Double.self, forKey: .longitude)
-        self.coordinate = CLLocationCoordinate2D(latitude: latitude, longitude: longitude)
-    }
-    
-    func encode(to encoder: Encoder) throws {
-        var container = encoder.container(keyedBy: CodingKeys.self)
-        try container.encode(coordinate.latitude, forKey: .latitude)
-        try container.encode(coordinate.longitude, forKey: .longitude)
-    }
-    
-    static func == (lhs: CoordinateWrapper, rhs: CoordinateWrapper) -> Bool {
-        lhs.coordinate.latitude == rhs.coordinate.latitude &&
-        lhs.coordinate.longitude == rhs.coordinate.longitude
-    }
-}
