+++ emoji-map/Views/Home/Home.swift
@@ -1,0 +1,18 @@
+//
+//  Home.swift
+//  emoji-map
+//
+//  Created by Enrique on 3/13/25.
+//
+
+import SwiftUI
+
+struct Home: View {
+    var body: some View {
+        Text(/*@START_MENU_TOKEN@*/"Hello, World!"/*@END_MENU_TOKEN@*/)
+    }
+}
+
+#Preview {
+    Home()
+}
